Error

See merge request webmobile2-sub3/s02p13b103!40

+++ fooddeuk/front/src/components/common/UserFeed.vue
@@ -1,4 +1,7 @@
 <template>
+<v-app data-app>
+
+
     <div class="wrapC">
         <div class="wrapper">
             <div class="profile-card js-profile-card">
@@ -19,14 +22,14 @@
                         <div class="profile-card-inf">
                             <div class="profile-card-inf__item">
                                 <div class="profile-card-inf__title">{{follower}}</div>
-                                <div class="profile-card-inf__txt">Followers</div>
+                                <div class="profile-card-inf__txt" style="color:black">Followers</div>
                             </div>
 
-                        <div class="profile-card-inf__item">
-                            <div class="profile-card-inf__title">{{following}}</div>
-                            <div class="profile-card-inf__txt">Following</div>
-                        </div>
-                    </div> 
+                            <div class="profile-card-inf__item">
+                                <div class="profile-card-inf__title">{{following}}</div>
+                                <div class="profile-card-inf__txt" style="color:black">Following</div>  
+                            </div>
+                        </div> 
                 </div>
             </div>
 
@@ -36,23 +39,35 @@
             </div>
             <div v-if="auth==0 || (auth==1 && isfollow==1)">
                 <div v-if="!post" style="margin-top:20px; text-align:center"> 게시물이 없습니다.</div>
-            
-                  <div v-for="(item,index) in list" v-bind:key="item.num">
-                    
+                  <div v-for="(item,index) in list" v-bind:key="item.num">  
                     <v-card
                             max-width="100%"
                             class="mx-auto"
-                            style="margin-bottom:100px"
+                            style="margin-bottom:100px; position:relative"
                     >
                     <v-list-item>
-                        <v-list-item-avatar><img src="../../assets/images/profile_default.png"></v-list-item-avatar>
+                        <v-list-item-avatar style="height:50px; width:50px"><img src="../../assets/images/profile_default.png"></v-list-item-avatar>
                         <v-list-item-content style="padding-left:5%">
-                        <v-list-item-title class="headline">{{item.title}}
-                        <button style="float:right" @click="updateFeed(item.num,item.title,item.content,item.count_star,item.address,item.image)">수정</button>
-                        <button style="float:right" @click="removeFeed(item.num)">삭제</button>
+                        <v-list-item-title style="margin-left:5px; margin-top:5px; font-size:15px;">{{item.title}}
+                            <v-menu offset-y style="float:right;">
+                            <template v-slot:activator="{ on }">
+                                <v-btn icon v-on="on" style="float:right">
+                                    <v-icon>mdi-dots-vertical</v-icon>
+                                </v-btn>
+                            </template>
+                            <v-list>
+                                <v-list-item>
+                                    <button style="float:right" @click="updateFeed(item.num,item.title,item.content,item.count_star,item.address,item.image)">수정</button>
+                                </v-list-item>
+                                <v-list-item>
+                                    <button style="float:right" @click="removeFeed(item.num)">삭제</button>
+                                </v-list-item>
+                            </v-list>
+                            </v-menu>
                         </v-list-item-title>
 
-                        <v-list-item-subtitle style="width:50px">{{nickname}} <div style="float:right"> {{getTime(item.date)}}</div> </v-list-item-subtitle>
+                        <v-list-item-subtitle style="width:50px; margin-left:5px">{{nickname}} <br>
+                         <div style="margin-top:10px; margin-left:2px"> {{getTime(item.date)}}</div> </v-list-item-subtitle>
                         <!-- <v-list-item-subtitle>{{getTime(item.date)}}</v-list-item-subtitle> -->
                         </v-list-item-content>
                         </v-list-item>
@@ -67,7 +82,8 @@
                             <br>
                             <v-card-text>
                                 {{item.content}}
-                                <img v-bind:src="item.image"  style="width:100%; heigh:auto; ">
+                                
+                                <img v-if="item.image!=='null'" v-bind:src="item.image"  style="width:100%; heigh:auto; ">
                             <br>
                             <br><br><hr><br>
                             주소 : {{item.address}} 
@@ -87,8 +103,18 @@
                         <button @click="commentview(item.num, index)"><img style="width:26px; margin-bottom:5px" src="../../assets/images/comment.png"></button>
                         </div>
                         <div style="width:33%; float:left; text-align:right; padding-right:10px; ; margin-top:3px">
-                        <button><img style="width:26px; margin-bottom:5px" src="../../assets/images/share.png"></button>
+                        <a href="#open-modal"><img style="width:26px; margin-bottom:5px" src="../../assets/images/share.png"></a>
+                        </div>
 
+                        <div id="open-modal" class="modal-window">
+                            <div style="height:250px; background-image: linear-gradient(to right,#7f53ac 0,#657ced 100%);">
+                                <h1 style="margin-top:20px">스크랩</h1>
+                                <v-text-field style="color:blue; width:90%" label="제목입력" v-model="scraptitle" id="scraptitle" hide-details="auto"></v-text-field>
+                                <v-text-field style="color:blue; width:90%" label="내용입력" v-model="scrapcontent" id="scrapcontent" hide-details="auto"></v-text-field>
+                                <a href="#">
+                                <button class="close-modal" style="margin-top:20px" @click="scrapfeed(item.num, scraptitle, scrapcontent); href='#'">제출</button>
+                                </a>    
+                            </div>
                         </div>
                         <br>
                     </div>
@@ -115,7 +141,7 @@
 
                         <div v-if="coment[index]===true">
                             <div v-for="cmt in todolist[index]" v-bind:key="cmt.id" >        
-                                <div style="margin-bottom:5px" v-for="cmts in cmt" v-bind:key="cmts.id" >
+                                <div style="margin-bottom:1px" v-for="cmts in cmt" v-bind:key="cmts.id" >
                                     <h5 style="float:left; margin-left:5px; margin-right:20px; font-weight:bold;"> {{ cmts.nickname }}</h5> &nbsp; 
                                     <h5 style="float:left; ">{{ cmts.comment }} 
                                     </h5>
@@ -142,6 +168,7 @@
             </div>
         </div>  
     </div> 
+    </v-app>
 </template>
 
 
@@ -179,7 +206,6 @@
             }
 
             this.nickname = this.propsNickname;
-
             //포스트 불러오기
             this.getUserByNickname(this.nickname);
 
@@ -213,6 +239,7 @@
                 this.isSubmit = isSubmit;
             },
             getTime(time) {
+                moment.locale('ko')
                 return moment(time).fromNow();
             },
              getUserByNickname(nick) {
@@ -225,7 +252,6 @@
                     this.intro = Response.data.intro;
                     this.email = Response.data.email;
                     this.auth = Response.data.auth;
-                
                     //팔로잉 내역 불러오기
                     this.getFollowing(this.email);
                     //팔로워 내역 불러오기
@@ -261,12 +287,11 @@
             getPostByNum(num) { //포스트가져오기
                 let form = new FormData()
                 form.append('num', num)
-                
+
                 http.get("/post/post/{num}?num="+num + '&email=' + this.$store.state.userinfo.email)
                 .then(Response => {
-                   
-                    this.post = Response.data.object; 
-               
+                    this.post = Response.data.object;
+          
                     //좋아요와 댓글 토글용 배열 생성
                     for (let index = 0; index < this.post.length; index++) {
                      
@@ -282,7 +307,7 @@
                         this.commentcount.push(this.post[index].count_comment)
                     }
 
-            
+                    
                 })
                 .catch(Error => {
                     console.log(Error)
@@ -352,7 +377,6 @@
                     
                     http.post("/follow/nonfollow", form)
                     .then(Response => {
-                        console.log(Response)
                         if(Response.data==='success') {
                             this.updateAlarmToFirebase();
                             alert("팔로우가 요청되었습니다.")
@@ -397,22 +421,20 @@
                 .catch(Error => {
                      console.log(Error)
                 })
-
             },
             commentview(num,index){ //댓글 버튼 누를 때
-          
             //댓글 불러오기
             if(this.coment[index]==false){
-                    http.get('/comment/comment?postnum='+num )
+                    http.get('/comment/comment?postnum='+num)
                 .then(response => {
-                    
                     if(response.data.object!=null){
                         this.todolist[index].push(response.data.object)
                     } 
-                    
+                
                    
                 })
-                .catch(Error =>{
+                .catch(Error => {
+                    console.log(Error)
                 })
 
             // //댓글 숨기기
@@ -433,10 +455,10 @@
                     .catch(Error => {
                         console.log(Error)
                     })
+                    
             }
             ,
             addcomment(num,index) {
-                
                 let form = new FormData()
                 form.append('comment', this.newcomment)
                 form.append('email', this.$store.state.userinfo.email)
@@ -453,7 +475,6 @@
                     .then(Response => {
                     
                         this.$set(this.commentcount,index,Response.data)
-       
                     })
                     .catch(Error => {
                         console.log(Error)
@@ -486,23 +507,29 @@
                 })
                 .catch(Error =>{
                 })
+ 
+                
             },
             
             removeFeed(num){
-                console.log(num)
-                 http.delete("/post/post?num=" + num + "&mynum=" + this.$store.state.userinfo.num)
-                .then(response => {
-                    alert('게시물이 삭제되었습니다.')
-                    console.log(response.data)
-                    this.post = response.data.object
-                    
-                })
+                 if (confirm("정말 삭제하시겠습니까??") == true){    //확인
+                    http.delete("/post/post?num=" + num + "&mynum=" + this.$store.state.userinfo.num)
+                    .then(response => {
+                        alert('게시물이 삭제되었습니다.')
+                        console.log(response.data)
+                        this.post = response.data.object
+                    })
                 .catch(Error =>{
                 })
+
+                }else{   //취소
+
+                    return false;
+
+                }
             },
             updateFeed(num, title, content, count_star, address, image){
                 var router = this.$router
-                console.log(image)
                  router.push({
                     name: "UpdateFeed",
                     params: {
@@ -596,15 +623,32 @@
                 })
                 alert("댓글이 등록되었습니다.")
                
+            },
+            scrapfeed(num,title,content) {
+                let form = new FormData()
+                form.append('postnum', num)
+                form.append('title',title)
+                form.append('content',content)
+                form.append('num',this.$store.state.userinfo.num)
+                http.post("/post/scrap", form)
+                .then(Response => {
+                })
+                .catch(Error => {
+                    console.log(Error)
+                })
             }
            
         },
         data: () => {
             return {
+                dialog: false,
+                usernum:0,
                 isSubmit: false,
                 error:{
                     comment:false
                 },
+                scraptitle:'',
+                scrapcontent:'',
                 nick:'',
                 nickname : '',
                 num:0,
@@ -625,12 +669,9 @@
                 userAlarmCount: 0,
                 likelist:[],
                 mynum:0,
-                commentcount:[],
-                
+                commentcount:[],         
                 //무한스크롤
                 list:[],
-                
-               
             }
         },
         components:{
@@ -639,12 +680,81 @@
         }
     }
 </script>
-<style>
-p {
+<style lang="scss" scoped>
+#app {
 
+    position:relative;
+    z-index:3
+}
+p {
     margin-left:5px;
+    margin-bottom:1px;
     color: gray;
     font-size:12px;
+}
+.modal-window {
+  text-align: center;
+  box-shadow: #aaa;
+  position: absolute;
+  background-color: rgba(255, 255, 255, 0.25);
+  top: 0;
+  right:0;
+  bottom: 0;
+  left: 0;
+  z-index: 999;
+  visibility: hidden;
+  opacity: 0;
+  pointer-events: none;
+  transition: all 0.3s;
+  &:target {
+    visibility: visible;
+    opacity: 1;
+    pointer-events: auto;
+  }
+  &>div {
+    background-color: ivory;
+    width: 300px;
+    position: absolute;
+    top: 50%;
+
+    background: #ffffff;
+  }
+  h1 {
+    font-size: 150%;
+    margin: 0 0 15px;
+  }
+}
+
+.modal-close {
+  color: rgb(0, 0, 0);
+  line-height: 50px;
+  font-size: 100%;
+  position: absolute;
+  right: 0;
+  text-align: center;
+  top: 0;
+  width: 70px;
+  text-decoration: none;
+  &:hover {
+    color: black;
+  }
+}
+.modal-window div:not(:last-of-type) {
+  margin-bottom: 15px;
+}
+
+small {
+  color: #aaa;
+}
+
+.btn {
+  background-color: #fff;
+  padding: 1em 1.5em;
+  border-radius: 3px;
+  text-decoration: none;
+  i {
+    padding-right: 0.3em;
+  }
 }
 
 </style>
